Type.Name instead of Type.CType (#45)
* Use Type.Name instead of Type.CType in SymbolTable

It seems like Type.Name is more common in gir. For instance,
implements refers to the interface by name and not c:type. This is also
handy because we do not have to strip off any trailing const or leading
pointers.

* Fix alias resolving

* Fix tests to use Type.Name

* Fix dealiasing

* Fix type lookup

* Void type is none

* Fix rebase

* Fix array type

Sometimes arrays do not have a name. In this case use Type.Name instead

* Fix GObject test

Object is now resolved as Class

* Remove void* test

* Remove unneeded method

* Rename back to AllRepositories

* Some cleanups and add some code comments